test(router): migrate
